Make functions out of methods that don't use `self`
