+++ pywizlight/scenes.py
@@ -34,3 +34,18 @@ SCENES = {
     32: "Steampunk",
     1000: "Rhythm",
 }
+
+
+def get_id_from_scene_name(scene: str) -> int:
+    """Return the id of an given scene name.
+
+    :param scene: Name of the scene
+    :type scene: str
+    :raises ValueError: Return if not in scene list
+    :return: ID of the scene
+    :rtype: int
+    """
+    for id in SCENES:
+        if SCENES[id] == scene:
+            return id
+    raise ValueError(f"Scene '{scene}' not in scene list.")
